Support loading model classes from across the CFS
Doctrine's default annotation driver does support a set of paths for
model/entity files, but as part of loading them it requires each file.
This breaks when using transparent extension, as there are potentially
multiple files in the include paths with duplicate class definitions.

The Doctrine_KohanaAnnotationDriver solves this by using a Kohana
approach to listing the model classes so model classes can live
anywhere within the CFS and support transparent extension in
the usual way.