Add support for doctrine/persistence >= v2.4.0
The 2.4.0 release had a behaviour change, in that
AnnotationDriver::setPaths() is now always called even if the
list of paths is empty.

That caused us to throw an exception as we have disabled support
for setting paths at all in favour of providing an explicit
list of entity classes.

Just ignore the call silently if the list is empty, to mirror
the old constructor behaviour.

https://github.com/doctrine/persistence/commit/6b12f2b4fd72328642515483f1b6913dbbd5d80a#diff-8032d1908cdfb96f4c529b4a3c641f21bfd8668774fabbf0c7c2491e1322e40eL85