Have NullPDO return a value for getAttribute(PDO::ATTR_SERVER_VERSION)
Doctrine needs to read a server version to create the database
platform driver, even if it is not actually checking database type
definitions. For example, in order to check class metadata.

The actual value is somewhat unimportant, so long as it is a valid
supported mysql version.

The outer connection class has the option to return a `null` if
the wrapped connection does not support getting a version, but
because we are using Doctrine's PDO wrapper that class indicates
that it does provide a version and can only return a string.